+++ app-android/src/main/java/com/wolfmontwe/mhp/challenge/mobile/android/app/domain/Result.kt
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2022. Wolf-Martell Montwé
+ */
+
+package com.wolfmontwe.mhp.challenge.mobile.android.app.domain
+
+import com.wolfmontwe.mhp.challenge.mobile.android.app.domain.Result.Failure
+import com.wolfmontwe.mhp.challenge.mobile.android.app.domain.Result.Success
+
+sealed interface Result<out R> {
+    data class Success<out T>(val data: T) : Result<T>
+    data class Failure(val exception: Exception) : Result<Nothing>
+}
+
+fun <T> Result<T>.isSuccess() = this is Success
+fun <T> Result<T>.isFailure() = this is Failure