+++ app-android/src/main/java/com/wolfmontwe/mhp/challenge/mobile/android/app/domain/entity/House.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2022. Wolf-Martell Montwé
+ */
+
+package com.wolfmontwe.mhp.challenge.mobile.android.app.domain.entity
+
+import com.wolfmontwe.mhp.challenge.mobile.android.app.domain.DomainContract.Entity
+
+data class House(
+    override val id: String
+) : Entity